new design proposal + detailed analysis
I wrote a new design proposal and detailed analysis of the options. The
goal is to complete several of the possible design in a single file
together with an example. This way we can start to make more informed
choices about what is a possibly good design

See the file `TypeclassBuilderEncodingMinimal.hs` for more information.