Monad laws hold true, type inference works.
So, I fixed the problems in TypeclassBuilderEncodingMinimal
by giving a bottom value to the second element in the
strict pair when no value is available. This way, the
Monoid restriction is removed, and type inference works as
it should.

The monad laws hold true now, and if we do not export
StrictPair, we can make sure the bottom value is never
evaluated.