Slight optimization for Monoid laws
Say `<> = mappend`. The Monoid laws say that

    x <> mempty == x

However, say `x = mempty`. This means
`x = Empty y`, and `x` could be the result of
`return y`.

The Monad laws dictate that

    return y >>= f == f y

Now, the above holds true for `return y`, but
recall the monoid law above. This would mean
that

    (return y <> mempty) >>= f == f y

should hold true as well. This is, however,
not the case, since `return y <> mempty` would
form an `Append` constructor without a value --
and thus, without a value, we cannot apply `f`.

This is now fixed by inspecting the uppermost
constructors before appending.