Remove the htmlNoDocType combinator
Rationale:

We should keep usage of the library as simple as
possible. Therefore, the `html` combinator should
do exactly what it says: insert an `<html>` tag.

Of course, it is convenient when the DocType is
inserted automatically, so I've added an
`docTypeHtml` combinator. The meaning of this
combinator should be immediately clear to the
user.