Fix unneeded qualified imports in website
The example code on the Blaze website had import statements along with
qualified import statements. This is unnecessary since if we just remove
the "qualified" keyword and import using an "as", we can achieve the
exact same effect (namely unqualified names when we want them and short
qualified names when we need them).